Move some helpers from alias_analysis to allocation_utils
For #4554
