+++ csrc/alias_analysis.cpp
@@ -17,7 +17,6 @@
 #include <ir/iostream.h>
 #include <ir/utils.h>
 #include <linked_hash_map.h>
-#include <logical_domain_map.h>
 #include <multidevice/utils.h>
 
 namespace nvfuser {
@@ -56,80 +55,6 @@ class AliasFinder : public OptOutConstDispatch {
   AliasAnalysisResult& analysis_;
 };
 
-// A helper function used to compute the perferred output layout. It computes
-// the mapping from `in_logical` to `out_root` and applies that mapping to
-// `preferred_in_layout`. For many ops, this function returns a good initial
-// preferred output layout for aliasing because it tries to preserve the input
-// layout. An op (e.g. ViewOp and SliceOp) that transforms root to logical
-// using expressions will have to modify this initial layout so its allocation
-// domain will be a function of its logical domain.
-//
-// Returns `nullopt` if computation fails, so the caller can handle things
-// conservatively.
-std::optional<Layout> mapInLayoutToOutRoot(
-    const std::optional<Layout>& preferred_in_layout,
-    TensorView* in,
-    TensorView* out) {
-  if (!preferred_in_layout.has_value()) {
-    return std::nullopt;
-  }
-
-  if (!ir_utils::computePermutation(
-           in->getLogicalDomain(), preferred_in_layout->allocation_domain)
-           .has_value()) {
-    // Give up when `in`'s allocation domain is not an logical permutation. As
-    // an extension, we could map in_alloc to in_logical and apply the inverse
-    // mapping to out_root.
-    return std::nullopt;
-  }
-
-  std::unordered_map<IterDomain*, IterDomain*> in_logical_to_out_root =
-      PairwiseLogicalDomainMap(in, out).mapProducerToConsumer();
-
-  Layout preferred_out_layout;
-  for (auto&& [in_alloc_id, contiguity] :
-       zip(preferred_in_layout->allocation_domain,
-           preferred_in_layout->contiguity)) {
-    IterDomain* out_root_id = getOrDefault(in_logical_to_out_root, in_alloc_id);
-    if (out_root_id == nullptr) {
-      // This can happen when in_alloc_id is of type reduction or squeezed out.
-      continue;
-    }
-    preferred_out_layout.allocation_domain.push_back(out_root_id);
-    preferred_out_layout.contiguity.push_back(contiguity);
-  }
-  return preferred_out_layout;
-}
-
-namespace {
-
-bool contiguityIsCompliant(
-    const std::optional<bool>& actual,
-    const std::optional<bool>& required) {
-  if (actual == true && required == false) {
-    return true;
-  }
-  return actual == required;
-}
-
-// Returns whether `layout` is compliant with `required`. This is
-// uni-directional. For example, `contiguity=[t,t]` is compliant with
-// `contiguity=[f,f]` but not vice versa.
-bool isCompliantWith(const Layout& layout, const Layout& required) {
-  if (layout.allocation_domain != required.allocation_domain) {
-    // This can be relaxed by allowing broadcast dimensions to be ordered
-    // differently.
-    return false;
-  }
-
-  for (const auto i : arange(layout.size())) {
-    if (!contiguityIsCompliant(layout.contiguity[i], required.contiguity[i])) {
-      return false;
-    }
-  }
-  return true;
-}
-
 bool okToRelayout(
     const TensorView* tv,
     const Layout& new_layout,
@@ -145,7 +70,6 @@ bool okToRelayout(
   }
   return isCompliantWith(new_layout, *old_layout);
 }
-} // namespace
 
 bool AliasFinder::aliasIfCompliant(
     const TensorView* alias,
